+++ anthem/__init__.py
@@ -2,7 +2,7 @@
 # Copyright 2016 Camptocamp SA
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl-3.0.en.html)
 
-__version__ = "0.4.0"
+__version__ = "0.5.0"
 
 # publish the decorator so we can use 'anthem.log'
 from output import log  # noqa